fix: use controlled accordion for labs as well
